requirements.txt
# Data collection
yfinance>=0.2.31
fredapi>=0.5.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0
requests>=2.31.0

# LLM providers
google-genai>=1.0.0
anthropic>=0.30.0

# Rendering
jinja2>=3.1.0
matplotlib>=3.8.0

# Email
resend>=0.7.0

# Google Sheets
gspread>=6.0.0
google-auth>=2.0.0

# Config
pyyaml>=6.0.0

# Testing
pytest>=8.0.0
pytest-mock>=3.12.0
